Move non UIA specific CustomObjects classes to Core
