+++ src/Core/CustomObjects/Converters/ElementTypeConverter.cs
@@ -1,17 +1,17 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Axe.Windows.Core.Bases;
 using System;
-using System.Globalization;
 
-namespace Axe.Windows.Desktop.UIAutomation.CustomObjects.Converters
+namespace Axe.Windows.Core.CustomObjects.Converters
 {
-    class BoolTypeConverter : ITypeConverter
+    class ElementTypeConverter : ITypeConverter
     {
         public string Render(dynamic value)
         {
             if (value == null) throw new ArgumentNullException(nameof(value));
-            return ((bool)value).ToString(CultureInfo.InvariantCulture);
+            return ((A11yElement)value).Glimpse;
         }
     }
 }
