+++ src/Core/CustomObjects/Converters/StringTypeConverter.cs
@@ -1,17 +1,16 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Axe.Windows.Core.Bases;
 using System;
 
-namespace Axe.Windows.Desktop.UIAutomation.CustomObjects.Converters
+namespace Axe.Windows.Core.CustomObjects.Converters
 {
-    class ElementTypeConverter : ITypeConverter
+    class StringTypeConverter : ITypeConverter
     {
         public string Render(dynamic value)
         {
             if (value == null) throw new ArgumentNullException(nameof(value));
-            return ((A11yElement)value).Glimpse;
+            return ((string)value);
         }
     }
 }
