+++ src/CoreTests/CustomObjects/Converters/IntConverterTests.cs
@@ -1,27 +1,26 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Axe.Windows.Desktop.UIAutomation.CustomObjects.Converters;
+using Axe.Windows.Core.CustomObjects.Converters;
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 
-namespace Axe.Windows.DesktopTests.UIAutomation.CustomObjects.Converters
+namespace Axe.Windows.CoreTests.CustomObjects.Converters
 {
     [TestClass()]
-    public class PointConverterTests
+    public class IntConverterTests
     {
         [TestMethod, Timeout(1000)]
         public void RenderTest()
         {
-            dynamic value=new double[2] { 42.0, 42.0 };
-            Assert.AreEqual("[x=42,y=42]", new PointTypeConverter().Render(value));
+            Assert.AreEqual("42", new IntTypeConverter().Render(42));
         }
 
         [TestMethod, Timeout(1000)]
         public void NullRenderTest()
         {
-            Assert.ThrowsException<ArgumentNullException>(() => new PointTypeConverter().Render(null));
+            Assert.ThrowsException<ArgumentNullException>(() => new IntTypeConverter().Render(null));
         }
     }
 }
